+++ java-client/src/main/java/co/elastic/clients/elasticsearch/core/ReindexRequest.java
@@ -45,6 +45,7 @@ import co.elastic.clients.util.ObjectBuilder;
 import co.elastic.clients.util.WithJsonObjectBuilderBase;
 import jakarta.json.stream.JsonGenerator;
 import java.lang.Boolean;
+import java.lang.Float;
 import java.lang.Long;
 import java.util.HashMap;
 import java.util.Map;
@@ -76,7 +77,7 @@ public class ReindexRequest extends RequestBase implements JsonpSerializable {
 	private final Boolean refresh;
 
 	@Nullable
-	private final Long requestsPerSecond;
+	private final Float requestsPerSecond;
 
 	@Nullable
 	private final Boolean requireAlias;
@@ -169,7 +170,7 @@ public class ReindexRequest extends RequestBase implements JsonpSerializable {
 	 * API name: {@code requests_per_second}
 	 */
 	@Nullable
-	public final Long requestsPerSecond() {
+	public final Float requestsPerSecond() {
 		return this.requestsPerSecond;
 	}
 
@@ -317,7 +318,7 @@ public class ReindexRequest extends RequestBase implements JsonpSerializable {
 		private Boolean refresh;
 
 		@Nullable
-		private Long requestsPerSecond;
+		private Float requestsPerSecond;
 
 		@Nullable
 		private Boolean requireAlias;
@@ -392,7 +393,7 @@ public class ReindexRequest extends RequestBase implements JsonpSerializable {
 		 * <p>
 		 * API name: {@code requests_per_second}
 		 */
-		public final Builder requestsPerSecond(@Nullable Long value) {
+		public final Builder requestsPerSecond(@Nullable Float value) {
 			this.requestsPerSecond = value;
 			return this;
 		}
