+++ java-client/src/main/java/co/elastic/clients/elasticsearch/core/ReindexRethrottleRequest.java
@@ -35,7 +35,7 @@ import co.elastic.clients.util.ApiTypeHelper;
 import co.elastic.clients.util.ObjectBuilder;
 import co.elastic.clients.util.ObjectBuilderBase;
 import jakarta.json.stream.JsonGenerator;
-import java.lang.Long;
+import java.lang.Float;
 import java.lang.String;
 import java.util.HashMap;
 import java.util.Map;
@@ -55,7 +55,7 @@ import javax.annotation.Nullable;
 
 public class ReindexRethrottleRequest extends RequestBase {
 	@Nullable
-	private final Long requestsPerSecond;
+	private final Float requestsPerSecond;
 
 	private final String taskId;
 
@@ -79,7 +79,7 @@ public class ReindexRethrottleRequest extends RequestBase {
 	 * API name: {@code requests_per_second}
 	 */
 	@Nullable
-	public final Long requestsPerSecond() {
+	public final Float requestsPerSecond() {
 		return this.requestsPerSecond;
 	}
 
@@ -100,7 +100,7 @@ public class ReindexRethrottleRequest extends RequestBase {
 
 	public static class Builder extends ObjectBuilderBase implements ObjectBuilder<ReindexRethrottleRequest> {
 		@Nullable
-		private Long requestsPerSecond;
+		private Float requestsPerSecond;
 
 		private String taskId;
 
@@ -110,7 +110,7 @@ public class ReindexRethrottleRequest extends RequestBase {
 		 * <p>
 		 * API name: {@code requests_per_second}
 		 */
-		public final Builder requestsPerSecond(@Nullable Long value) {
+		public final Builder requestsPerSecond(@Nullable Float value) {
 			this.requestsPerSecond = value;
 			return this;
 		}
