+++ java-client/src/main/java/co/elastic/clients/elasticsearch/core/UpdateByQueryRequest.java
@@ -48,6 +48,7 @@ import co.elastic.clients.util.ObjectBuilder;
 import co.elastic.clients.util.WithJsonObjectBuilderBase;
 import jakarta.json.stream.JsonGenerator;
 import java.lang.Boolean;
+import java.lang.Float;
 import java.lang.Long;
 import java.lang.String;
 import java.util.HashMap;
@@ -119,7 +120,7 @@ public class UpdateByQueryRequest extends RequestBase implements JsonpSerializab
 	private final Boolean requestCache;
 
 	@Nullable
-	private final Long requestsPerSecond;
+	private final Float requestsPerSecond;
 
 	@Nullable
 	private final String routing;
@@ -391,7 +392,7 @@ public class UpdateByQueryRequest extends RequestBase implements JsonpSerializab
 	 * API name: {@code requests_per_second}
 	 */
 	@Nullable
-	public final Long requestsPerSecond() {
+	public final Float requestsPerSecond() {
 		return this.requestsPerSecond;
 	}
 
@@ -657,7 +658,7 @@ public class UpdateByQueryRequest extends RequestBase implements JsonpSerializab
 		private Boolean requestCache;
 
 		@Nullable
-		private Long requestsPerSecond;
+		private Float requestsPerSecond;
 
 		@Nullable
 		private String routing;
@@ -924,7 +925,7 @@ public class UpdateByQueryRequest extends RequestBase implements JsonpSerializab
 		 * <p>
 		 * API name: {@code requests_per_second}
 		 */
-		public final Builder requestsPerSecond(@Nullable Long value) {
+		public final Builder requestsPerSecond(@Nullable Float value) {
 			this.requestsPerSecond = value;
 			return this;
 		}
